+++ src/NexNet.IntegrationTests/NexNet.IntegrationTests.csproj
@@ -11,7 +11,6 @@
   <ItemGroup>
     <Compile Remove="AsyncAutoResetEventTests.cs" />
     <Compile Remove="CountingResetAwaiterTests.cs" />
-    <Compile Remove="NexusClientTests_NexusDuplexPipe.cs" />
   </ItemGroup>
 
   <ItemGroup>
